utils/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
altair==4.1.0
argon2-cffi @ file:///D:/bld/argon2-cffi_1602546829425/work
asgiref==3.3.1
astor==0.8.1
astroid @ file:///C:/ci/astroid_1592481955828/work
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backports.functools-lru-cache==1.6.1
base58==2.1.0
beautifulsoup4==4.9.3
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
blinker==1.4
bokeh==2.3.1
boto3==1.17.47
botocore==1.20.47
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1602537445122/work
chardet @ file:///D:/bld/chardet_1602255574834/work
click==7.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
configparser==5.0.2
cryptography @ file:///D:/bld/cryptography_1604179260895/work
cycler==0.10.0
Cython==0.29.14
dataclasses==0.6
datasets==1.5.0
datefinder==0.7.1
datefinder-lexpredict==0.6.2.1
dateparser==0.7.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
discord.py==1.6.0
Django==3.1.3
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
filelock==3.0.12
fsspec==0.9.0
future==0.18.2
gast==0.3.3
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.23.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.15
grpcio==1.33.2
h5py==2.10.0
hatesonar==0.0.7
html5lib==1.1
huggingface-hub==0.0.8
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
image==1.5.33
imagesize==1.2.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1602263269022/work
ipykernel @ file:///D:/bld/ipykernel_1602683107803/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
jellyfish==0.6.1
Jinja2==2.11.2
jmespath==0.10.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1598486169312/work
jupyter-core @ file:///D:/bld/jupyter_core_1605128035695/work
jupyterlab==2.2.9
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1593951277307/work
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lexnlp==1.8.0
libcst==0.3.13
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.3
MarkupSafe @ file:///D:/bld/markupsafe_1602267515311/work
matplotlib==3.3.3
mccabe==0.6.1
mistune @ file:///D:/bld/mistune_1605115874185/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
multiprocess==0.70.11.1
mypy-extensions==0.4.3
mysql==0.0.2
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1602859080374/work
nbconvert @ file:///D:/bld/nbconvert_1603792834939/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1602732862338/work
nbmultitask==0.1.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1604290829246/work
nltk==3.5
notebook @ file:///D:/bld/notebook_1605103885768/work
num2words==0.5.10
numba==0.53.1
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.2.4
pandocfilters==1.4.2
pathtools==0.1.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///C:/ci/pillow_1603821929285/work
plot-model==0.20
prettytable==2.0.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
promise==2.3
prompt-toolkit==1.0.15
proto-plus==1.11.0
protobuf==3.13.0
psutil==5.8.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydeck==0.6.1
pydot==1.4.1
pydot-ng==2.0.0
pydotplus @ file:///tmp/build/80754af9/pydotplus_1594390100614/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1603558917696/work
pylint @ file:///C:/ci/pylint_1598605958085/work
PyNaCl==1.4.0
pynndescent==0.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent @ file:///D:/bld/pyrsistent_1605115843590/work
PySocks @ file:///D:/bld/pysocks_1602327025506/work
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2020.4
pywin32==228
pywinpty @ file:///D:/bld/pywinpty_1602377770565/work
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.7.14
remote-pdb==2.1.0
reporters-db==2.0.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.6
sacremoses==0.0.44
scikit-learn==0.23.1
scikit-surprise==1.1.1
scipy==1.5.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.0.0
seqeval==1.2.2
shortuuid==1.0.1
simplegeneric==0.8.1
simpletransformers==0.61.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sklearn==0.0
smart-open==3.0.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.1
streamlit==0.79.0
subprocess32==3.5.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado @ file:///D:/bld/terminado_1605117153893/work
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.11.1
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado @ file:///D:/bld/tornado_1604105307492/work
tqdm==4.51.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
transformers==4.5.0
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
typing-inspect==0.6.0
tzlocal==2.1
umap==0.1.1
umap-learn==0.5.1
Unidecode==1.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1603125704209/work
us==2.0.2
validators==0.18.2
wandb==0.10.25
watchdog==2.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
whatsapp-web==0.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1602260090416/work
wincertstore==0.2
wordcloud==1.8.1
wordninja==2.0.0
wrapt==1.12.1
xgboost==1.3.3
xlrd==2.0.1
xxhash==2.0.0
yarl==1.6.3
youtube-dl @ file:///D:/bld/youtube-dl_1618611559431/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
